refac: Moved UpdateDownloader.cs
Moved this class to Install namespace
