+++ ListenToIt.Core/ListenToIt.Updater/ListenToIt.Updater.csproj
@@ -64,11 +64,11 @@
     <Compile Include="CmdOptions\InstallOptions.cs" />
     <Compile Include="CmdOptions\UpdateOptions.cs" />
     <Compile Include="Install\Installer.cs" />
+    <Compile Include="Install\UpdateDownloader.cs" />
     <Compile Include="Package\PackageInfo.cs" />
     <Compile Include="Package\Version.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="UpdateDownloader.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
